Delete role endpoint

Approved-by: Nung Khual

+++ controller/roleController.js
@@ -1,6 +1,7 @@
 const Joi = require("joi");
 Joi.objectId = require("joi-objectid")(Joi);
 const Role = require("../model/Role");
+const Admin = require("../model/Admin");
 const Permission = require("../model/Permission");
 
 //@desc  Create a role
@@ -216,8 +217,14 @@ const editRole = async (req, res) => {
 const deleteRole = async (req, res) => {
   try {
     if (req.admin?.permissions.includes("admin_delete")) {
+      const admin = await Admin.findOne({ role: req.params.id });
+      if (admin) {
+        return res.status(200).send({
+          message: "Cannot delete as this role is assign to some admin.",
+        });
+      }
       await Role.deleteOne({ _id: req.params.id });
-      res.status(200).send({ message: "Role deleted successfully" });
+      return res.status(200).send({ message: "Role deleted successfully" });
     }
     return res.status(200).send({ error: "Not allowed to delete role" });
   } catch (err) {
